+++ metadata.rb
@@ -1,19 +1,25 @@
-name             'elixir'
-maintainer       'Chef Software, Inc.'
-maintainer_email 'cookbooks@chef.io'
-license          'Apache-2.0'
-description      'Manages an Elixir installation'
-version          '1.1.0'
+name              'elixir'
+maintainer        'Sous Chefs'
+maintainer_email  'help@sous-chefs.org'
+license           'Apache-2.0'
+description       'Manages an Elixir installation'
+version           '1.1.0'
+source_url        'https://github.com/sous-chefs/elixir'
+issues_url        'https://github.com/sous-chefs/elixir/issues'
+chef_version      '>= 12.15'
 
-%w(ubuntu debian redhat centos suse opensuse opensuseleap scientific oracle amazon).each do |os|
-  supports os
-end
+supports 'amazon'
+supports 'centos'
+supports 'debian'
+supports 'opensuseleap'
+supports 'oracle'
+supports 'redhat'
+supports 'scientific'
+supports 'suse'
+supports 'ubuntu'
 
 depends 'git', '>= 5.0.0'
 depends 'github'
 depends 'erlang', '>= 3.0.0'
 depends 'libarchive', '>= 2.0.0'
 
-source_url 'https://github.com/chef-cookbooks/elixir'
-issues_url 'https://github.com/chef-cookbooks/elixir/issues'
-chef_version '>= 12.15'